% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_info}
\alias{get_data_info}
\title{Get information about a data frame in the environment}
\usage{
get_data_info(data_name)
}
\arguments{
\item{data_name}{Name of the data frame as character string}
}
\value{
List containing data information or NULL if not found
}
\description{
Get information about a data frame in the environment
}
\examples{
get_data_info("mtcars")
}
