% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_available_data}
\alias{list_available_data}
\title{List all available data frames in the environment}
\usage{
list_available_data()
}
\value{
Character vector of data frame names
}
\description{
List all available data frames in the environment
}
\examples{
list_available_data()
}
