% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_execute_plot}
\alias{safe_execute_plot}
\title{Safely execute generated R plotting code}
\usage{
safe_execute_plot(code, data_env = .GlobalEnv, verbose = FALSE)
}
\arguments{
\item{code}{Character string containing R code}

\item{data_env}{Environment containing the data (default: .GlobalEnv)}

\item{verbose}{If TRUE, print the code being executed}
}
\value{
Logical TRUE if successful, FALSE otherwise
}
\description{
Safely execute generated R plotting code
}
